PIM-3725: move object id resolver and fix repository interfaces issues
