+++ src/Pim/Bundle/CatalogBundle/Doctrine/ArrayApcCache.php
@@ -9,6 +9,8 @@ use Doctrine\Common\Cache\Cache;
 /**
  * Proxy for ApcCache in http mode and ArrayCache in command line mode
  *
+ * @deprecated will be removed in 1.4
+ *
  * @author    Antoine Guigan <antoine@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
