+++ src/Pim/Bundle/CatalogBundle/Repository/ChannelRepositoryInterface.php
@@ -37,6 +37,13 @@ interface ChannelRepositoryInterface extends IdentifiableObjectRepositoryInterfa
     public function getDeletedLocaleIdsForChannel(ChannelInterface $channel);
 
     /**
+     * Return an array of channels
+     *
+     * @return array
+     */
+    public function getChannels();
+
+    /**
      * Return an array of channel codes
      *
      * @return array
