--- src/Pim/Bundle/CatalogBundle/Resources/config/resolvers.yml
@@ -1,13 +1,0 @@
-parameters:
-    pim_catalog.resolver.object_id.class: Pim\Bundle\CatalogBundle\Doctrine\Common\ObjectIdResolver
-
-services:
-    pim_catalog.resolver.object_id:
-        class: %pim_catalog.resolver.object_id.class%
-        arguments:
-            - '@pim_catalog.doctrine.smart_manager_registry'
-        calls:
-            - [ addFieldMapping, ['family', %pim_catalog.entity.family.class%]]
-            - [ addFieldMapping, ['option', %pim_catalog.entity.attribute_option.class%]]
-            - [ addFieldMapping, ['category', %pim_catalog.entity.category.class%]]
-            - [ addFieldMapping, ['group', %pim_catalog.entity.group.class%]]
