+++ src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/orm.yml
@@ -101,7 +101,7 @@ services:
     pim_catalog.doctrine.query.filter.family:
         class: %pim_catalog.doctrine.query.filter.family.class%
         arguments:
-            - '@pim_catalog.resolver.object_id'
+            - '@pim_catalog.doctrine.query.filter.object_id_resolver'
             - ['family.id', 'family.code']
             - ['IN', 'NOT IN', 'EMPTY']
         tags:
@@ -110,7 +110,7 @@ services:
     pim_catalog.doctrine.query.filter.groups:
         class: %pim_catalog.doctrine.query.filter.groups.class%
         arguments:
-            - '@pim_catalog.resolver.object_id'
+            - '@pim_catalog.doctrine.query.filter.object_id_resolver'
             - ['groups.id', 'groups.code']
             - ['IN', 'NOT IN', 'EMPTY']
         tags:
